+++ .gitignore
@@ -14,3 +14,4 @@ Cargo.lock
 
 /__pycache__/
 rust-toolchain
+result
