+++ Cargo.lock.copy.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: 2022 Felix Robles <felix@sequentech.io>
+
+SPDX-License-Identifier: AGPL-3.0-only
