+++ src/wasm.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2022 Felix Robles <felix@sequentech.io>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+#[cfg(feature = "wasmtest")]
+pub mod wasm_ballot;
